Refactor auth types and imports for better type separation
